update: undo changes to aws_config::connector module remove: impl Default for HttpConnector
